Allows configurable number of entries in feed
Addresses issue #1. By adding a `?length=XXX` parameter to the URL, the
number of entries in the feed can be configured to be between 1 and 100
(a limit imposed by tumblr2rss). I set 100 as the max because A) it
already takes quite a long time to round-trip to tumblr and B) because
I'm worried tumblr will rate limit me.